+++ scripts/nd2_describe.py
@@ -9,7 +9,7 @@ from dataclasses import asdict
 from pathlib import Path
 
 import nd2
-from nd2.readers._chunk_decode import iter_chunks
+from nd2._parse._chunk_decode import iter_chunks
 
 
 def _get_version(path):
