+++ src/nd2/_parse/_chunk_decode.py
