+++ src/nd2/_parse/_clx_lite.py
