+++ src/nd2/_parse/_legacy_xml.py
